Clear ALL Compiler Warnings (must accompany core PR) (#2470)
<!--
Thank you for contributing to mod-playerbots, please make sure that
you...
1. Submit your PR to the test-staging branch, not master.
2. Read the guidelines below before submitting.
3. Don't delete parts of this template.

DESIGN PHILOSOPHY: We prioritize STABILITY, PERFORMANCE, AND
PREDICTABILITY over behavioral realism.

Every action and decision executes PER BOT AND PER TRIGGER. Small
increases in logic complexity scale
poorly across thousands of bots and negatively affect all. We prioritize
a stable system over a smarter
one. Bots don't need to behave perfectly; believable behavior is the
goal, not human simulation.
Default behavior must be cheap in processing; expensive behavior must be
opt-in.

Before submitting, make sure your changes aligns with these principles.
-->

<!-- Describe what this change does and why it is needed -->
Followup to several warnings cleanups done few months ago by
@Celandriel. The hope is that we can live in a world where our children
never see a compiler warning on the master branch.

PR is in fourcommits.

1. The first one is obvious incorrect casts, reordering of header
signatures, incorrect brackets.

2. The second deletes unused variables/code and redundant user-defined
copy constructor.

3. The third removes EquipmentSlots enum declaration, and it must be
merged with this core commit
https://github.com/mod-playerbots/azerothcore-wotlk/pull/222

Four years ago there was a crash happening in RandomItemMgr.h because of
the declaration of EquipmentSlots.
https://github.com/NoxMax/azerothcore-wotlk/commit/db71f4739cd2c3d3c5ffa8d17cc2aa5481c1650d
The correct fix was to include Player.h. Instead the enum was
forward-declared in Playerbots, and changed in Player.h to uint32. That
fixed the crash, but left a warning ever since. The third commit is a
proper fix to the crash, that reverts the core changes, and doesn't
create warnings.

4. Last one modifies the PR template to remind people not to add new
warnings. I think we should eventually edit the .yml file for GH Action
to treat warnings as errors, but for now, I am just adding this as an
advice for the short term before we enforce it. I pasted the checkbox in
this PR description too just as an immediate example.

<!--
If your PR is very minimal (comment typo, wrong ID reference, etc), and
it is very obvious it will not have
any impact on performance, you may skip these question. If necessary, a
maintainer may ask you for them later.
-->

<!-- Please answer the following: -->
- Describe the **minimum logic** required to achieve the intended
behavior.
- Describe the **processing cost** when this logic executes across many
bots.

<!--
- Step-by-step instructions to test the change.
- Any required setup (e.g. multiple players, number of bots, specific
configuration).
- Expected behavior and how to verify it.
-->

Changes should compile without any errors and warnings, if you apply it
to the current state of test-staging. There might be other errors
introduced from accompanying new PRs.

<!-- As a generic test, before and after measure of pmon (playerbot pmon
tick) can help you here. -->
- Does this change increase per-bot/per-tick processing or risk scaling
poorly with thousands of bots?
    - - [x] No, not at all
    - - [ ] Minimal impact (**explain below**)
    - - [ ] Moderate impact (**explain below**)

- Does this change modify default bot behavior?
    - - [x] No
    - - [ ] Yes (**explain why**)

- Does this change add new decision branches or increase maintenance
complexity?
    - - [x] No
    - - [ ] Yes (**explain below**)

<!--
AI assistance is allowed, but all submitted code must be fully
understood, reviewed, and owned by the contributor.
We expect contributors to be honest about what they do and do not
understand.
-->
Was AI assistance used while working on this change?
- - [ ] No
- - [x] Yes (**explain below**)
<!--
If yes, please specify:
- Purpose of usage (e.g. brainstorming, refactoring, documentation, code
generation).
- Which parts of the change were influenced or generated, and whether it
was thoroughly reviewed.
-->
Had Claude go through a long, long, list of warnings of all those
int/uint comparisons.

<!--
TRANSLATIONS:
Anything new that the bots say in chat must be in a translatable format.
This is done using GetBotTextOrDefault,
which you can search for in the codebase to find examples. Your code
needs to have English as the default fallback,
while the full translations need to be in an SQL update file. The
languages in the file are the nine language
options supported by AzerothCore: English, Korean, French, German,
Chinese, Taiwanese, Spanish, Spanish Mexico, and
Russian. See
data/sql/playerbots/updates/2025_12_27_ai_playerbot_fishing_text.sql as
an example of a translation SQL
update, whose content are called within the codebase at
src/strategy/actions/FishingAction.cpp
-->

- - [x] Stability is not compromised.
- - [x] Performance impact is understood, tested, and acceptable.
- - [x] Added logic complexity is justified and explained.
- - [x] Any new bot dialogue lines are translated.
- - [x] Documentation updated if needed (Conf comments, WiKi commands).
- - [x] New and modified files do not introduce new compiler warnings.

<!-- Anything else that's helpful to review or test your pull request.
-->

Do not fear the 70 file diff. The 70 file diff should fear you.
But seriously it's not as scary as it looks.

---------

+++ src/Ai/Base/Actions/InventoryAction.cpp
@@ -401,7 +401,7 @@ ItemIds InventoryAction::FindOutfitItems(std::string const name)
 std::string const InventoryAction::parseOutfitName(std::string const outfit)
 {
     uint32 pos = outfit.find("=");
-    if (pos == -1)
+    if (pos == uint32(-1))
         return "";
 
     return outfit.substr(0, pos);
@@ -415,7 +415,7 @@ ItemIds InventoryAction::parseOutfitItems(std::string const text)
     while (pos < text.size())
     {
         uint32 endPos = text.find(',', pos);
-        if (endPos == -1)
+        if (endPos == uint32(-1))
             endPos = text.size();
 
         std::string const idC = text.substr(pos, endPos - pos);
